backend/hr_service/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlmodel>=0.0.14
psycopg2-binary>=2.9.9
redis>=5.0.1
python-jose[cryptography]>=3.3.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2
fakeredis>=2.20.1
python-multipart>=0.0.6
reportlab>=4.0.7
celery>=5.3.4
python-dateutil>=2.8.2
openpyxl>=3.1.2
python-magic>=0.4.27